refactor: split server.ts into route modules and utility modules
- Extract game event processing into src/game-event.ts
- Extract sound utilities into src/sounds.ts
- Create src/routes/mappings.ts for mapping.json CRUD and user-sounds
- Create src/routes/sounds.ts for sound file management
- Create src/routes/discord.ts for discord member endpoint
- Slim server.ts to thin orchestrator (~60 lines)
- Add @/ path alias and update all imports
- Add sounds/ to gitignore
