+++ .gitignore
@@ -19,3 +19,4 @@ build/
 
 mapping.json
 settings.json
+sounds/
